refactor(frontend): move support and admin utility stores to runes
